Split monolithic admin.module.ts into controller/service/module files
